Extract generic DDS fuzz model from interval collections fuzz tests (#14675)
## Description

This separates the portions of the interval collection fuzz tests which
are DDS-specific from those that are SharedString/interval collection
specific. The goal of this refactor is to enable moving the generic DDS
harness to a shared location (I'm currently thinking
`@fluid-internal/test-dds-utils` is a good fit) to reduce boilerplate in
other DDS's fuzz models.

It also includes some mild refactors & feature additions to
stochastic-test-utils, which were implemented as needed (or as I found
the current API overly restrictive).