r11s: add initial summary as valid parent (#14772)
## Description

Continuation of #14401.

There is an additional scenario where client summarizer can propose a
valid parent summary, but Scribe rejects it because it does not meet
scribe's conditions.

1. Initial Summary - "abcd"
2. Some ops
    1. Join
    2. Leave
    3. NoClient
4. NoClient op triggers Service Summary to start
5. New client joins, loading from initial summary "abcd"
6. Service Summary completes - "1234"
7. Client proposes summary with parent "abcd" (is actually valid)
8. Scribe has no knowledge of initial summary version, "abcd", as valid
parent
9. Scribe rejects "abcd" for not being equal to "1234"

The client recovers from this by retrieving the latest service summary
and retrying, but it is inefficient and not correct from client
perspective. Also, it can cause E2E tests to fail.

## Fix

Add the versionId of the initial summary as a valid parent summary
within the DB's Scribe Checkpoint during document creation. Then, when
scribe boots for a new document, it has knowledge of initial summary as
a valid parent.

Importantly, do not set `lastClientSummaryHead` to the initial summary
version because it was not technically a client summary, and therefore
has no Ack.

## Repro

1. Add an artificial delay in Scribe's service summary write to simulate
slow prod environment
1. Load fluid app
2. immediately refresh app
3. wait for summary proposal to be nacked