Partial Implementation of Representing Muted Changes in Optional Fields (#14747)
## Description
This PR adds a representation for changes made to deleted nodes. It's
only complete to the point that a node exists constraint could be
written that will work in a unit test but not an end to end test.
There's an included skipped end to end test that isn't working yet
properly as we're not sure how to go about representing intermediate
changes in composition.

This PR adds a new contract with regard to representation as follows:

If the `deletedBy` property is set in an OptionalFieldChange,
`childChange` refers to the node deleted in the field in the revision
given by `deletedBy`. Otherwise, `childChange` refers to the node
currently present in the field.