Simplify SharedTree indexes (#14865)
## Description

This PR strips away much of the "index" types/code in order to simplify
the relationship between SharedTreeCore and SharedTree. It is not clear
exactly what an index is or what an index's requirements are at this
time, and the current and planned features that are considered
"index-like" differ in enough ways that there is not much obvious value
in trying to conform them to a common interface or pattern. After this
PR, the process for organizing and exposing indexes is simply "have
SharedTree do it however it wants". In the future, we may want to create
an index abstraction to help guide external implementers to make their
own indexes, but it isn't clear at this time the best way to do that so
for now we can just keep things simple.

* Removed generic index parameter from SharedTreeCore
* Replaced *Index classes with *Summarizer classes. They are now only
responsible for summarizing their respective state; they do not do any
event registration.
* Simplified the way that SharedTrees provide their summarizers to
SharedTreeCore.
* IndexEvents are now "ChangeEvents" and are available to subclasses of
SharedTreeCore.
* EditManagerSummarizer is a concern of SharedTreeCore, so it has been
moved to shared-tree-core rather than being awkwardly injected in by
SharedTree.