Discard GC data when loading state from snapshot with different GC version than current (#14864)
When GC loads from a snapshot or updates latest state from a snapshot
(in refreshLatestSummary), if the GC version of the snapshot is
different from the current GC version, reset all local GC state and only
update deleted nodes from this snapshot.

When the GC version of a snapshot is different from the current, the GC
data in the snapshot cannot be trusted. The GC data has to be
re-generated (in the next GC run) because the previous GC data was
generated by a version that may have issues - this already happens
today. The same logic should apply when updating the garbage collector's
state from the snapshot. Note that we still update the deleted nodes
state because these nodes have been deleted and cannot be brought back.


[AB#3779](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3779)