ADO#3900 - TestObjectProvider.loadContainer will default to use delay load mode
Currently default initial connection mode is "write" when loading a
container. However, if delayed load mode is use, the connection mod is
"read".  We don't have test cover when the container is loaded as "read"
by default.

This work item will partially make our e2e cover delayed mode
because of the default connection mode behavior. This is part one of the
work that will cover tests that use TestObjectProvider.  Following work
item ADO#3901 will track the rest.

The only test that needs to be fixed are AgentScheduler, where it try to
use register a task to generate an 'op' to force it to switch to write
mode. However, the op is not generate if the task is already there.  The
fix is to generate different name every time `forceWrite` is called.