Ignore old snapshots and assume gc should be regenerated (#14906)
From Hammon Pang
> It looks like there are some older files (last updated in 2022) that
were used recently but can’t summarize due to the following
FluidElementNotFound error during PostSnapshot:
>
> Cannot locate node with path '[omitted]/gc' under '[omitted]'.

The solution here is that for older snapshots gc will now ignore, and
regenerate the whole gc blob and submit it. The thought here is that we
cannot use older snapshots for doing incremental GC summary. So, to keep
things simple we ignore them and regenerate the GC data. This is
expensive but it should happen very rarely.

---------

Co-authored-by: Navin Agarwal <45832642+agarwal-navin@users.noreply.github.com>