Fix for deleted nodes changes not considered when doing incremental summary of GC data (#14914)
## Bug
If the only change in GC data is deleted nodes, the GC data is written
as a handle and the changes to deleted nodes will be lost. The issue is
the following code which does not compare deleted nodes across
summaries:
```
    // If nothing changed since last summary, send a summary handle for the entire GC data.
    if (
        this.latestSummaryData.serializedGCState === serializedGCState &&
        this.latestSummaryData.serializedTombstones === serializedTombstones
    ) {
        const stats = mergeStats();
        stats.handleNodeCount++;
        return {
            summary: {
                type: SummaryType.Handle,
                handle: `/${gcTreeKey}`,
                handleType: SummaryType.Tree,
            },
            stats,
        };
    }

```
Since sweep is not enabled yet, deleted nodes are not written so this
should not be a problem yet.

## Fix
The fix is to compare deleted nodes across summaries as well in the
above piece of code. Also, added unit tests for incremental summaries.


[AB#3903](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3903)