Fix debug logger on error output in tests (#14949)
`DebugLogger` create a bound function of `console.error` to log error.
In our e2e tests, we overwritten the function to do nothing so we don't
create noise in tests. However, that also means that depending on when
the `DebugLogger` is created, the debug output will not be output to the
console even if the debug namespace is enabled.

The change is to check if the namespace is enabled, and use the default
logging function it is enabled.