Improve Time to Debug for Debug Current Mocha Test (#14953)
## Description
Simplify our Lauch.json configuration for Debug Current Mocha Test. This
has improved time to debug, and doesn't seem to introduce any new
issues. The previous config was put together to work around out of
memory issues hit in the vs code debugger which don't appear to occur
anymore. This change also seems to be inline with the current guidance
for debugging typescript in the vs code documentation:
https://code.visualstudio.com/docs/typescript/typescript-debugging

Co-authored-by: Tony Murphy <anthonm@microsoft.com>