Change `TestObjectProvider.loadContainer` default to use delayed load mode (#14916)
Currently when loading a container, the default for the initial
connection is "write" mode. However, if delayed load mode is used, the
initial connection mode is "read". We don't have test coverage when the
container is loaded as "read" by default.

This PR fixes
[AB#3900](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3900).
It will partially make our e2e test to cover delayed mode because of the
default connection mode behavior. This is part one of the work that will
cover tests that use `TestObjectProvider.loadContainer`. A future PR for
the next work item
[AB#3904](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3904)
and
[AB#3901](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3901)
will track the rest.

The change add the `loadMode.deltaConnection = "delayed"` to the header,
if the caller doesn't specify and `loadMode` header. The exception is
that if the test asked for `reconnect: false` in the header, we can't
use delayed load mode, because the `connect` call implicitly assumes we
want to reset the reconnection mode to enabled, which is a possible API
gap.

The only tests that needed to be fixed are `AgentScheduler` tests, where
it tries to use the call to register a task to generate an 'op' and
force it to switch to write mode. However, the op is not generated if
the task already exists. The fix is to generate different name every
time `forceWrite` is called.

The change is also been tested with targeting ODSP, and no additional
unique failures compare to baseline in the `main` branch.