Add attribution to experimental map (#14952)
## Description

As a follow-up of the
https://github.com/microsoft/FluidFramework/pull/13989, this PR includes
the integration of the original SharedMap and its attribution-related
API, along with the creation of a new experimental DDS named
`attributableMap`. This experimental DDS allows for the tracking of
attribution information, such as the user who made an update and the
timestamp of the change. The `attributableMap` currently includes three
types of attribution keys: Op-stream, detached, and local.

- The Op-stream attribution key indicates that the update of attribution
information is synchronized with the processing of a map operation, and
only occurs once the update is ack'd.
- The detached attribution key only occurs when the DDS is not attached,
and can be summarized into a snapshot.
- The local attribution key exists when the operation has not yet been
ack'd and cannot be summarized.

### Work breakdown

- Attribution API implementation:
eddd6892502fd6234425417b58faecbbdae41cb3
- Mocha test suites: 3ca07e3da0334179fd59772e9acb9e5dd49c4747
- API document: 7976aa8fab0ea313b3ac0c61eb861485e7bd73be and
b6631ef3715fadb47911e20a938b1cb62a4c3581

### PR Checklist

-   [X] I have updated the documentation accordingly.
-   [X] I have added tests to cover my changes.
-   [X] All new and existing tests passed.
-   [X] My code follows the code style of this project.
- [X] I ran the lint checks which produced no new errors nor warnings
for my changes.
- [X] I have checked to ensure there aren't other open Pull Requests for
the same update/change.

### Does this introduce a breaking change?

-   [ ] Yes
-   [X] No