Add v1 identifier index to SharedTree (#14934)
## Description

This adds a new index to SharedTree: the identifier index. It scans the
tree for valid identifiers that are under a special identifier trait,
and places them in an index for quick lookup. A user can then pass an
identifier to the index to quickly retrieve the EditableTreeNode that
has that index.

Things to add in future versions:
* Use runtime provided compressed IDs for Identifiers rather than just
"number"
* A more efficient method of scanning the tree for IDs. Currently the
entire tree is rescanned after every change. Instead, it could just
respond to deltas or tree events and make only the minimal changes to
the identifier map.
* A more efficient data structure for storing the mappings from
identifier to node. Right now, it is just a hashmap, which is quick for
lookup but very slow to clone (which happens when forking
SharedTreeViews)
* A way to configure what the identifier trait is within SharedTree. The
trait is already configurable with respect to the index, but
`SharedTree` currently has it hardcoded
* More documentation. The API is likely to change as related parts of
the system evolve, so a fully-featured doc experience for using the
identifier index is still pending