refactor(client-debugger): Refactor types to better represent hierarchy of devtools (#14957)
The purpose of this PR is to prepare the devtools system for tooling
related to Fluid runtime concepts that exist outside of the scope of the
Fluid Container. It refactors existing types to better encapsulate
scoped responsiblities.

At a high level the change looks like the following:
* "Registry" -> "FluidDevtools"
* Originally, this object strictly existed to enumerate Container-level
devtools (previously "debugger") instances so consumers could listen for
changes to the list of registered Container tooling instances. But going
forward this root level object will be the manager of devtools-related
tracking of metadata that lives outside of the scope of an individual
Container (e.g. telemetry, network traffic / errors, etc.).
* This PR does not change what data is managed by the root devtools
object, but semantically prepares for such changes in the future.
* "FluidClientDebugger" -> "ContainerDevtools"
* This is mostly just a rename, since we are going with "devtools" for
this tooling (instead of "debugger") going forward.

Other changes include:
* No longer storing a static singleton of the root devtools on the
Window (globalThis). Instead, applications will be responsible for
storing and managing the lifetime of this object. A static singleton
made sense for the "registry" usecase, but not for something like the
root devtools, which are semantically bound to the lifetime of the
application.
* Rename some other related "debugger" members to use "devtools"
terminology.
* Simplify some event names and update docs.

---------

Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>