[GC stress]: Extend summarizer cache expiry timeout to 5 mins (#14989)
Extending the summarizer cache expiry timeout to 5 mins. This means that
the summarizer will load from cache more often. This is important
because this flow is a major source of bugs in GC data. When summarizer
loads from a cached snapshot and has to update state from a newer
snapshot, the summary / GC state has to be updated and this flow has
known to have bugs. This change will potentially repro existing or new
bugs.
We are also considering making big changes to this flow to address these
bugs - let's get them reproing in the stress tests and then see them
fixed.


[AB#3963](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3963)