[GC stress] Log reference events to telemetry and remove console logs (#15022)
Removed most of the console logs from the GC stress tests. The ones ones
reamaining are when a data store / attachment blob is referenced /
unreferened. These are also logged to kusto so that when bugs are found,
we can build a timeline of what happeneded w.r.t. references.


[AB#2138](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2138)

##Reviewer guidance
This is for an experimental stress test in test/gc-stress branch. This
is not in the main FF branch and does not affect its stress test