[GC stress] Add not changing data store / DDS  (#15023)
Add not changing a data store / DDS for a while. This will give us
coverage when a data store / DDS is unchanged across summaries or
unchanged after loading simulating virtualization scenarios. For
example, when refreshing from snapshot, if a data store is unchanged, it
goes down a different path and ends up updating the latest summary state
of its children later on.
Currently, the data stores continuously send ops so there is rarely a
summary in which data stores do not change.