[GC stress] Use real data store and blob ids in test to correate logs (#15067)
Use the data store / attachment blob's read id in the stress test. These
are logged to telemetry when a data store / blob is referenced or
unreferenced. Using the read ids will help correlate them in telemetry.

## Reviewer guidance
This is for an experimental stress test in test/gc-stress branch. This
is not in the main FF branch and does not affect its stress test