[GC stress] Limit the number of attachment blobs that can be retrieved (#15073)
Currently, clients can upload and download unlimited number of
attachment blobs causing throttling. This change limits the number of
attachment blobs that can be uploaded / downloaded to 6 per client. This
should reduce the number of network calls.

## Reviewer guidance
This is for an experimental stress test in test/gc-stress branch. This
is not in the main FF branch and does not affect its stress test.