Correctly document interval endpoints as exclusive (#15050)
## Description

I previously documented interval endpoints as being inclusive on both
ends. Though the end endpoint is *placed* directly on the character
position, this yields behavior consistent with half-open intervals.

This PR updates documentation in the README/public API surface to be
more in-line with how we currently recommend its usage, and calls out
some gotchas (like the fact that in its current form, applications using
intervals probably want to add an "end of string" marker to their
SharedString)

Note: the upcoming interval stickiness feature will likely change some
of the recommendations here (perhaps including APIs we expose, hopefully
in a way that simplifies them). But best to fix it up for now to
recommend best practices for people using it.