MergeTree Revertible: Remove reliance on RevertibleDriver for Append (#15065)
## Description
With the changes made in #13301 it possible to access the merge tree
itself during the append phase without the revertible driver. This
change does not modify the public interface, but modifies the internals
of revertibles to leverage the merge tree directly. This work is a
precursor to reusing merge revertibles in matrix, which has a
significantly different architecture which makes accessing a driver in
the append phase difficult, without significantly altering the public
interfaces of the package.

Additionally, this change will allow the simplification of the revert
driver itself.


related to
[AB#1880](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/1880)

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>