fix(build-cli): Use shell in exec and bump (#15117)
When running in CI and in some local environments, the `flub exec` and
`flub bump` commands don't seem to work with some commands, such as
`npm`, because they can't be found by the spawned subprocess. I worked
around this by passing `shell: true` into the options, but that's not a
great long-term solution. Longer term solution is tracked by
[AB#4067](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4067).

I also updated the execa usage in the `bump` command to pass arguments
using an array. I originally did this for debugging but it seems more
flexible and reliable than the old string-only approach, so I left it in
place.