[GC stress] Combine data store and blob activity type generation (#15122)
Currently, the activity types for data stores and blobs are
independently generated at random. Each has a None activity type where
no activity happens. The goal of the None activity is to have summaries
where the data store containing references does not changee. However,
since the two activitiy types are independently generated, the chances
of getting a None event are very rare. This change combines the activity
type generation which increases the chances of None activity for a data
store to 25%.

## Reviewer guidance
This is for an experimental stress test in test/gc-stress branch. This
is not in the main FF branch and does not affect its stress tes