refactor (protocol-definitions): Deprecate MessageType.RemoteHelp Op (#15127)
https://github.com/microsoft/FluidFramework/issues/9709

Deprecating MessageType.RemoteHelp. I have confirmed it's a concept that
no longer exists and that none of the server teams are using this code
any longer, so I am deprecating it in this PR, with a follow up remove
PR to come. There is code in the foreman lambda that uses this Op. I
have confirmed that it is okay to remove that code also. A different PR
will remove the foreman lambda code since it is a different release
group.