Update SharedTree branch APIs to be more flexible (#15111)
## Description

This PR adds capabilities to the branch and view/fork APIs so that they
can merge and rebase with more flexibility. Specifically, they are no
longer restricted to pulling from, rebasing onto, and merging into their
"parent" (the branch from which they were forked), but can do these
operations with any other branch/view/fork. Additional details:

* Doing a merge no longer automatically pulls/rebases onto the branch
being merged into, and the method direction has been inverted (you call
"me.merge(you)" now rather than "me.merge()" - which would have been
equivalent to "you.merge(me)"). This gives more flexibility to the APIs,
and also means that for both the rebase and merge operations, only the
branch on which the method is invoked is mutated, never the argument
branch. If a user still wants the old behavior of doing a "pull" before
a "merge", they can simply call `b.rebaseOnto(a)` before `a.merge(b)`.
* The `AnchorSet` parameter when constructing a branch has been made
optional. If passed, the branch will manage the rebasing of the anchor
set, but it is not required. Since the ownership of anchor sets is in
flux and not clear, this should provide flexibility for different
decisions in the future.
* A suite of unit tests for branches has been added
* The unit tests for SharedTree views and transactions have been cleaned
up
* Branch class code has been cleaned up

## Breaking Changes

`ISharedTreeFork.pull()` and `ISharedTreeFork.merge()` have been
replaced with `ISharedTreeView.merge()` and
`ISharedTreeFork.rebaseOnto()`