ci(client): Run webpack in CI test phase (#13895)
When we switched to pnpm for package management, one thing that we broke
accidentally was some webpack builds. Part of the reason is that we
don't run webpack in CI! This PR corrects that by adding webpack as an
additional test task.

This does add ~90-120 seconds to a CI build, but #13894 will switch most
of them to use esbuild instead, which will negate most of the added
time.