feat(api-markdown-documenter): Add support for multiple package entry-points (#15150)
For packages with a single entry-point, the generated output structure
remains unchanged. For packages with multiple entry points, an extra
document will be generated for each entry-point under the package
directory.
- Note that we still will only generate at most 1 document for each API
item, regardless of how many entry-points in which they appear.

Includes a unit test for the scenario.

Note: API-Extractor doesn't currently support this, but it is planned,
and the API report format supports it.

### Breaking Change

These changes constitute a breaking change due to changing
transformation config options (removes `transformApiPackage` and adds
`transformApiEntryPoint`).