Update getIntervalId return type and intervalCollection test setup  (#15168)
Related to https://github.com/jzaffiro/FluidFramework/pull/3. Intervals
now have immutable IDs, so there's no need to return undefined from this
API anymore. I had this in the linked PR, but it was decided that the
change should go into main instead of the interval topic branch. While I
was writing tests for the linked item, I saw that the same sharedString
was accidentally being used to create two interval collections that
should be distinct.