Remove IDocumentAttributes.term and ISequenceDocumentMessage.term (again) (#15166)
These members were related to an experimental feature that did not ship. As a result they are unused/ignored by all consumers. They were deprecated in main in #14971, so this change removes them entirely.

It also removes the leaderEpoch concept from r11s which is related. This was an internal concept though, so no breaking change required.

To ensure consistency through the transition (as this is spread across multiple release groups), we can treat "term" as 1 in all cases, and "epoch" as 0 in all cases. These are already the values that we should be seeing in all cases, since we never increment them on the server anyway. Then as the new versions of packages are picked up, we can just eliminate these as well.

This is a second version of the change, after #14959 had to be reverted.