Fix ODSP Fetch 412 in Summarizer E2E test:  Simply close the summarizer just before uploading. (#15130)
## Description

Closing after the upload would not work on ODSP as the summarizer handle
would be returned, and the server would have this handle id stored as
the last summarizer. The next summarizer, getting from cache, would
fail.
We could have force to get from network, which would work, or we can
simply forcibly fail to upload by closing the summarizer instead. Fix
[#AB4075](https://dev.azure.com/fluidframework/internal/_workitems/edit/4075)