build: Add dependency guidelines to package READMEs (#15191)
Now that we are using `~` interdependencies within the repo, we want to
make sure that customers know that they should continue to use `^`
dependencies when depending on Fluid Framework packages. This PR adds a
new markdown-magic script and updates READMEs across the client release
group accordingly.

BREAKING CHANGE: This is technically a breaking change since the default
output for some functions changed.