Implement client joins in the dds fuzz harness (#15104)
## Description

This implements a technique @jatgarg used for the directory fuzz tests
in the generic dds fuzz harness enabling fuzz testing of clients joining
from intermediate summaries.

Since summarization must occur on a client without edits and the mocks
used for these tests don't spin up any of the associated summarizer
infrastructure, this puts an explicit summarizer client in the fuzz
test's state. Client joins are implemented by taking a summary and
loading from it. Since this is asynchronous, this forces usages to
convert to a bunch of async code.

I verified that existing usages (interval collection, task manager)
still report errors as expected on eventual consistency issues (this is
important as it can be easy to accidentally introduce problems by
failing to await a promise when adding asynchronicity).

TaskManager and interval collections both pass with this feature enabled
(after one small bugfix to interval collections I've added a regression
test for), so I went ahead and did so.