Add TestTreeProviderLite (#15200)
## Description

This replaces many of the SharedTree tests that used `TestTreeProvider`
with ones that use a new class `TestTreeProviderLite` instead.
`TestTreeProviderLite` uses minimal mocks to test a SharedTree, and is
suitable for tests that want to test a single tree or a few connected
trees. It does not allow for any control over summarization or
connection/reconnection flows like `TestTreeProvider` does, but it is
far lighter weight. It should be the default for tests that want to test
SharedTree, and `TestTreeProvider` should only be used for more
complicated scenarios that need the extra features. Some migrated tests
now run noticeably faster, for example the op size summary tests.