Rename @fluid-internal/tree to @fluid-experimental/tree2 (#15025)
## Description

Renames the `@fluid-internal/tree` package to
`@fluid-experimental/tree2`. This will make it included in releases
while clearly marking it as experimental. This change is simply being
done so contributors to the tree package can have packages builds to
test with.

Currently all public releases (internal, LTS, prerelease and test)
include the same set of packages, which includes `@fluid-experimental`
but not `@fluid-internal/tree`, so this change results in the package
being included in these builds.

All APIs in `@fluid-experimental/tree2` are marked as `@alpha`. This is
redundant with them being in an experimental package: both mean no API
stability commitment is being made and no data format compatibility
commitment is being made.

`pnpm run layer-check --md` was run to update PACAKGES.md. Looks like it
was out of date so some other updates are included.

## Breaking Changes

Uses of the previously non-public `@fluid-internal/tree` package will
need to update to `@fluid-experimental/tree2`