Fix assert tagging for experimental (#15215)
## Description

Before this change, the `experimental` directory was not processed by
assert tagging.

This change fixes several issues that prevent assert tagging for working
on `experimental`.

A future run of assert tagging will do the actual tagging for the
experimental directory.

There are two main issues this fixed:
- experimental/dds/attributable-map/src/mapKernel.ts had short codes
which collided with those in packages/dds/map/src/mapKernel.ts
- experimental/dds/tree had its own `assert` method with a signature
that did not work with tagging. It has been renamed to
`assertWithMessage` and usages have either been ported to a different
assert, or moved to the renamed assertWithMessage.