Connect in write mode if Container is dirty (#14997)
## Description
This fixes a bug when loading with stashed ops with `deltaConnection:
"none"` headers. After `connect()` is called, instead of connecting in
write mode, Container would attempt to connect in read mode, but for
reasons I don't fully understand, would get stuck in `CatchingUp` and
never transition to `Connected`. Runtime should at all times correctly
reflect when there are or are not pending stashed ops via the dirty
signal, and Container should connect in write mode when it is true.

Related to #14502