Fix a bug where child summarizer node's paths are incorrectly set to parent's paths (#15222)
## Bug
When a child summarizer node is created, if a parent node has any
pending summaries, it creates one for the child node as well. In [this
instance](https://github.com/microsoft/FluidFramework/blob/4f62d86b9b0469a133414bb2feae8f7c929f60fb/packages/runtime/container-runtime/src/summary/summarizerNode/summarizerNode.ts#L680),
the path of the parent node is incorrectly copied over to the child
summarizer node.
Once this happens, if the child node does not change until the next
summary, it's SummaryHandle's path will incorrectly point to the
parent's summary tree from the previous summary. The summary for this
node will now be corrupted. Any containers loading from this summary
cannot load this node anymore.
We see this happening in production for DDSs (child node = DDS, parent
node = data store).

Note that this is a rare scenario where a data store has to be realized
after a summary op is submitted and before the ack is received. That's
why we don't see a lot of hits for this.


[AB#4132](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4132)

## Fix
When creating pending summaries for child summarizer node, use the
child's latest summary's path.