Prevent SharedTree trunk commits from being evicted if they are referenced by branches (#15177)
## Description

This PR updates the EditManager's trunk eviction code to retain all
commits on the trunk that are currently the base for any outstanding
branch/fork. It attempts to do so while preserving
separation-of-concerns by adding additional events to the Branch class,
but not putting any trunk-specific code into the Branch class (e.g.
sequence numbers). It also:

* Deletes the EditManager module in Core and moves EditManager to be
part of SharedTreeCore. This is done since EditManager now has a
dependency on Branch, which would otherwise introduce a cyclic
dependency. Regardless, it is probably an improvement to the module
structure even without the motivation provided by this PR.
* Made some methods on SharedTreeCore protected instead of public to
reduce the API surface for eventual/future subclassers of SharedTreeCore
(e.g. "other SharedTrees"). These methods are now tested via a test
subclass in shared-tree-core/utils.ts