R11s: Simplify and share throttle config helpers (#15221)
## Description

Currently, the configs we have to set for Throttling in R11s are usually
unnecessarily complex. We can hide a lot of this complexity for the
typical use-cases, while still allowing overrides for specific needs to
keep the current level of flexibility and control.

### Typical Scenario

**Request:**
We can only support 10,000 HTTP requests per hour. Let's set our
throttle limits accordingly.

**Dev Work Today:**
If we want to check throttling on a 30 second interval, then for 10,000
limit per hour, then we can allow `(10,000 requests / 1h) * (1h / 60min)
* (1min / 60s) = 10,000 requests / 3600 seconds = 2.77 reqs/second`. So
then burst is `2.77 * 30 = 83.33 reqs/30s` and then perMs is `(2.77
requests / second) * (1s / 1000ms) = 0.00277 reqs/ms` and cooldown
should just equal throttle interval, and max cache age should be 2x
interval duration...
```js
{
  maxPerMs: 0.00277,
  maxBurst: 83.33,
  minCooldownIntervalInMs: 30_000,
  minThrottleIntervalInMs: 30_000,
  maxInMemoryCacheSize: 1_000,
  maxInMemoryCacheAgeInMs: 60_000,
  enableEnhancedTelemetry: false,
}
```

**Dev Work Now:**
If we want to check throttling on a 30 second interval, then for 10,000
limit per hour, then we can allow `(10,000 requests / 1h) * (1h / 60min)
* (1min / 60s) = 10,000 requests / 3600 seconds = 2.77 reqs/second`. So
then burst is `2.77 * 30 = 83.33 reqs/30s`.
```js
{
  maxPerInterval: 83.33,
  intervalInMs: 30_000
}
```

### Notes

This only cuts out 2 of the math steps, but it will make the configs
much more intuitive and maintainable, while still allowing the more
fine-tuned approach we have now if needed.

## Alternative Approach

We could even further reduce the maths needed by providing 3 options for
simplified config like
```ts
{
  maxPerIntervalDuration: 10_000,
  intervalDurationInMs: 3_600_000, // ms in 1 hour
  throttleCheckIntervalInMs: 30_000,
}
```
where `throttleCheckIntervalInMs` would be the `30s` from above, and
then `maxPerIntervalDuration` would be relative to
`intervalDurationInMs`, and then internal maths would be done to convert
that to be relative to do the previously manual `a reqs / b time -> x
reqs / y seconds` conversion.

I think the downside of this alternative approach is that it still has
the `throttleCheckIntervalInMs` implementation detail, whereas the
approach I used in this PR is more "implementation agnostic" so that we
could keep these configs for different throttling implementations with a
different conversion helper in the code.